+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Bundler 1.12+ way of preserving ENV, use 1.0 for older bundler and 0.x.d for even older ones
 - Never run if Bundler not defined
+- No more tampering with exception if debundle failed
 
 ## 1.0.0